Forward Firewall: when resetting firewall, an error was generated in log, because there's no reread file

+++ html/cgi-bin/forwardfw.cgi
@@ -1997,7 +1997,9 @@ sub rules
 sub reread_rules
 {
 	system("/usr/local/bin/forwardfwctrl");
-	system("rm ${General::swroot}/forward/reread");
+	if ( -f "${General::swroot}/forward/reread"){
+		system("rm ${General::swroot}/forward/reread");
+	}
 }
 &Header::closebigbox();
 &Header::closepage();
